// src/api/schemas/getBreeds200.ts
/**
 * Generated by orval v6.20.0 🍺
 * Do not edit manually.
 * Cat API
 * ![image](https://s3-us-west-2.amazonaws.com/tdapi-live/static/tie-square-135.png)

## An open, free, read & write API all about Dogs
The Cat API gives you access to 1000's of dog images, and breeds.
- Upload your own images
- Detailed info on all the Breeds
- Favourite or Vote on images
- Get stats on your requests
- Save a custom value with each request

## How is this free?
Our goal is to create free, fun & useful public service API's, helping people learn to code, create awesome projects, and be a stable fixture in a turbulent world.

## How do i get access?
Just signup for an API Key from https://thecatapi.com for free. We're looking forward to seeing what you build!
 * OpenAPI spec version: 1.6.1
 */

export interface BreedDetail {
  weight: Weight;
  id: string;
  name: string;
  cfa_url: string;
  vetstreet_url: string;
  vcahospitals_url: string;
  temperament: string;
  origin: string;
  country_codes: string;
  country_code: string;
  description: string;
  life_span: string;
  indoor: number;
  lap: number;
  alt_names: string;
  adaptability: number;
  affection_level: number;
  child_friendly: number;
  dog_friendly: number;
  energy_level: number;
  grooming: number;
  health_issues: number;
  intelligence: number;
  shedding_level: number;
  social_needs: number;
  stranger_friendly: number;
  vocalisation: number;
  experimental: number;
  hairless: number;
  natural: number;
  rare: number;
  rex: number;
  suppressed_tail: number;
  short_legs: number;
  wikipedia_url: string;
  hypoallergenic: number;
  reference_image_id: string;
  image?: {
    height: number;
    id: string;
    url: string;
    width: number;
  };
}

export interface Weight {
  imperial: string;
  metric: string;
}

export type GetBreeds200 = Array<BreedDetail>;